Support sending large amounts of random data.
This commit adds support for sending large amounts of PRNG
data using chunked Transfer-Encoding.

Note the change in the specification of the request's body.
Previously, body was just a string.  Now the following format
is used:

  "body": {
    "content": <string>,
    "size": <number>,
    "type": "content|random",
  }

The old format is still supported for compatibility, but will
be removed in future versions.

A typical use to send PNRG data in the request's body is:

  "body": {
    "size": <number>,
    "type": "random",
  }